ecoScripts/aurora/Image/requirements.txt
kafka-python
pyyaml
requests
pymongo
